pyproject.toml
[tool.poetry]
name = "audep"
version = "0.1.0"
description = "automated deployment tool"
authors = ["coolxv <coolxv@gmail.com>"]
license = "MIT"

#packages = [
#    { include = "audep" },
#    { include = "extra_package/**/*.py" },
#]

[tool.poetry.dependencies]
python = "^3.6"
fabric2 = "^2.5"
toml =  "^0.10"
jmespath = "^0.9"
Jinja2 = "^2.11"
pyexpander = "^1.9"
lark-parser = "^0.8"
pluginbase = "^1.0"
glom = "^19.10"
pyyaml = "^5.3"
yamlordereddictloader = "^0.4"
chardet = "^3.0"
xmltodict = "^0.12"
apscheduler = "^3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyinstaller = "^3.6"


[[tool.poetry.source]]
name = "ten"
url = "https://mirrors.cloud.tencent.com/pypi/simple"

#[[tool.poetry.source]]
#name = "163"
#url = "https://mirrors.163.com/pypi/simple/"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
audep = "audep.cli:main"